requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Redis and Celery
redis==5.0.1
celery[redis]==5.3.6

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Google APIs
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.116.0

# AI and LangChain
openai==1.12.0
langchain==0.1.6
langchain-openai==0.0.5
langchain-community==0.0.19
faiss-cpu==1.13.2
tiktoken==0.5.2

# Utilities
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
httpx==0.26.0
structlog==24.1.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
black==24.1.1
ruff==0.2.0
mypy==1.8.0
